stupid tests to make sure guids are different each time

+++ AbanLibTests/facility/Test.cs
@@ -8,18 +8,20 @@ public abstract class Test : IDisposable
 
 	protected Test()
 	{
+		PreviousGuid = LatestGuid;
+		LatestGuid = GuidString;
 		DirectoryName = $"AbLibTest-{GuidString}";
 		var tempDirectory = Path.GetTempPath();
 		DirectoryPath = Path.Combine(tempDirectory, DirectoryName);
 		Directory.CreateDirectory(DirectoryPath);
 	}
 
+	protected static string PreviousGuid { get; private set; } = "";
+	protected static string LatestGuid { get; private set; } = "";
+
 	public virtual void Dispose()
 	{
 		GC.SuppressFinalize(this);
-		if (Directory.Exists(DirectoryPath))
-		{
-			Directory.Delete(DirectoryPath, true);
-		}
+		if (Directory.Exists(DirectoryPath)) Directory.Delete(DirectoryPath, true);
 	}
 }
